requirements.txt
fastapi
bittensor
sqlalchemy
python-dotenv